+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Allow `craft.formie.renderJs` to set JS attributes for scripts.
 - Added “Required Field Indicator” for forms, to either show an asterisk for required fields (default) or show optional for non-required fields.
 - Added `data-fui-alert-error` and `data-fui-alert-success` attributes on front-end alerts.
+- Added the `form.setPageSettings()` function to override page settings in your Twig templates.
 
 ### Changed
 - Compatible with Craft `5.0.0-beta.1`.